restructure & feature  - see details...
restructure - T.modeChangeCallbacks array now used for modules to
register interest in changes of the mode (i.e. drift on/off).
feature - speed hist now renders for drift too...the slivers of each bar
are coloured acording to the time of the pos sample contributing to that
bit of the bar..with the pos samples sorted within the bar.